Move unit tests into test/unit
