+++ test/unit/HeapJobOrderer.test.ts
@@ -1,8 +1,8 @@
 import { strict as assert } from 'assert'
-import { NormalJob } from '../src/Job/NormalJob'
-import { HeapJobOrderer } from '../src/JobOrderer/HeapJobOrderer'
-import { Job } from '../src/Job/Job'
-import { UnknownJobError } from '../src/JobOrderer/UnknownJobError'
+import { NormalJob } from '../../src/Job/NormalJob'
+import { HeapJobOrderer } from '../../src/JobOrderer/HeapJobOrderer'
+import { Job } from '../../src/Job/Job'
+import { UnknownJobError } from '../../src/JobOrderer/UnknownJobError'
 
 describe('HeapJobOrderer', () => {
 	it('correctly orders jobs', () => {
