+++ test/unit/ZeroconfClient.test.ts
@@ -2,9 +2,9 @@ import { mock } from 'jest-mock-extended'
 import { Bonjour, Browser, RemoteService } from 'bonjour'
 import { EventEmitter } from 'events'
 
-import { supplyClient } from '../src/Zeroconf/ZeroconfClient'
-import { Client } from '../src/Client/Client'
-import { SERVICE_TYPE } from '../src/Constants'
+import { supplyClient } from '../../src/Zeroconf/ZeroconfClient'
+import { Client } from '../../src/Client/Client'
+import { SERVICE_TYPE } from '../../src/Constants'
 
 const MOCK_SERVICE_HOST = 'example.com'
 const MOCK_SERVICE_PORT = 1337
